tests passing for simple goto definition
