+++ crates/lexing/src/lexed.rs
@@ -61,6 +61,12 @@ pub struct SyntaxKindInfo {
     pub position: Position,
 }
 
+impl SyntaxKindInfo { 
+  pub fn is_at_offset(&self, offset: u32) -> bool {
+    self.annotation <= offset && offset <= self.token
+  }
+}
+
 #[derive(Debug, Clone)]
 pub struct LexerError {
     message: Arc<str>,
@@ -176,6 +182,15 @@ impl Lexed<'_> {
         let high = self.infos[index].token as usize;
 
         &self.source[low..high]
+    }
+    pub fn text_at_offset(&self, offset: u32) -> Option<&str> {
+
+        let index = self.infos.iter().position(|info| info.is_at_offset(offset))?;
+
+        let low = self.infos[index].qualifier as usize;
+        let high = self.infos[index].token as usize;
+
+        Some(&self.source[low..high])
     }
 
     pub fn text_in_range(&self, range: Range<usize>) -> &str {
