+++ crates/lsp/tests/fixtures/goto_definition/01_simple.purs
@@ -1,0 +1,9 @@
+module Test where
+
+test1 = 1
+
+a = test1
+
+-- TEST 5:8 => 3:1-3:10
+-- TEST 5:9 => 3:1-3:10
+-- TEST 5:5 => 3:1-3:10
